LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-liquid-glass-positioning-expected.txt

Testing that MacOSInlineMediaControls uses CSS flexbox positioning.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaControls.rightContainer.element.style.left is ''
PASS mediaControls.rightContainer.element.style.left is ''
PASS successfullyParsed is true

TEST COMPLETE
